refactor: extract formatErrorMessage into common utility function
